+++ database/migrations/2025_04_24_144038_create_properties_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('category_id')->nullable()->constrained('categories_properties');
             $table->text('description');
             $table->integer('price');
-            $table->integer('square_meters');
+            $table->float('square_meters');
             $table->text('address')->nullable();
             $table->string('city')->nullable();
             $table->string('country')->default('Portugal');
